+++ EssentialFeed/EssentialFeediOS/Feed Image Loader/FeedImageDataLoader.swift
@@ -1,0 +1,17 @@
+//
+//  FeedImageDataLoader.swift
+//  EssentialFeediOS
+//
+//  Created by Alex Tapia on 10/10/21.
+//
+
+import Foundation
+
+public protocol FeedImageDataLoaderTask {
+  func cancel()
+}
+
+public protocol FeedImageDataLoader {
+  typealias Result = Swift.Result<Data, Error>
+  func loadImageData(from url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoaderTask
+}
